+++ ynab_sdk/utils/clients/cached_client.py
@@ -6,16 +6,26 @@ from redis import Redis
 from ynab_sdk.utils.clients.base_client import BaseClient
 from ynab_sdk.utils.configurations.cached import CachedConfig
 
+from typing import Union
+
 
 class CachedClient(BaseClient):
 
-    def __init__(self, config: CachedConfig):
+    def __init__(self, config: CachedConfig, TTL_value: Union[int, None] = 3600):
         super().__init__(config)
         self.redis = Redis(host=config.redis_host, port=config.redis_port, db=config.redis_db)
-        if config.redis_TTL and config.redis_TTL > 0:
-            self.redis_TTL = config.redis_TTL
+        self._redis_TTL = TTL_value
+
+    @property
+    def redis_TTL(self) -> Union[int, None]:
+        return self._redis_TTL
+
+    @redis_TTL.setter
+    def redis_TTL(self, TTL_value: Union[int, None]) -> None:
+        if TTL_value and TTL_value > 0:
+            self._redis_TTL = TTL_value
         else:
-            self.redis_TTL = None
+            self._redis_TTL = None
 
     def get(self, endpoint: str):
         self.logger.error(f'Endpoint => {endpoint}')
@@ -31,7 +41,7 @@ class CachedClient(BaseClient):
             response = requests.get(url, headers=self.headers)
             data = response.json()
             if response.status_code == 200:
-                self.redis.set(endpoint, json.dumps(data), self.redis_TTL)
+                self.redis.set(endpoint, json.dumps(data), self._redis_TTL)
             else:
                 self.logger.error(f'Error when getting {url}')
                 self.logger.error(data)
