+++ ynab_sdk/utils/configurations/cached.py
@@ -9,12 +9,3 @@ class CachedConfig(DefaultConfig):
         self.redis_host = redis_host
         self.redis_port = redis_port
         self.redis_db = redis_db
-        self._redis_TTL: Union[int, None] = 3600
-
-    @property
-    def redis_TTL(self) -> Union[int, None]:
-        return self._redis_TTL
-
-    @redis_TTL.setter
-    def redis_TTL(self, TTL_value: Union[int, None]) -> None:
-        self._redis_TTL = TTL_value
